backend/requirements.txt
# Backend — Fase 1+ (API, SQLAlchemy, Alembic, JWT)
fastapi>=0.115.0,<0.116
python-multipart>=0.0.9,<0.1
uvicorn[standard]>=0.32.0,<0.35
pydantic[email]>=2.10.0,<3
pydantic-settings>=2.6.0,<3

# Persistência (DEC-18)
sqlalchemy>=2.0.36,<2.1
psycopg[binary]>=3.2.0,<4
alembic>=1.14.0,<2

# Auth (DEC-16 — access JWT; refresh em evolução)
python-jose[cryptography]>=3.3.0,<4
bcrypt>=4.2.0,<5

# Testes e qualidade
httpx>=0.28.0,<0.29
pytest>=8.3.0,<9
pytest-cov>=6.0.0,<7
pytest-html>=4.1.0,<5
ruff>=0.8.0,<0.10
